[1126]: 회원가입 구현
- 회원가입 상호작용을 하나의 씬으로 통일
- RegisterDTO 수정
