+++ Assets/Scripts/Web/DTO/RegisterDTO.cs
@@ -1,18 +1,20 @@
+using Data;
+
 namespace Web.DTO
 {
     public class RegisterDTO : DTO
     {
-        public string userName { get; set; }
+        public string username { get; set; }
         public string password { get; set; }
-        public string userEmail { get; set; }
-        public string userNickname { get; set; }
+        public string user_nickname { get; set; }
+        public CharacterType character { get; set; }
 
-        public RegisterDTO(string userName, string password, string userNickname)
+        public RegisterDTO(string username, string password, string user_nickname, CharacterType character)
         {
-            this.userName = userName;
             this.password = password;
-            this.userEmail = "Email@gmail.com";
-            this.userNickname = userNickname;
+            this.username = username;
+            this.user_nickname = user_nickname;
+            this.character = character;
         }
     }
 }
